backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI Services (Lightweight API clients - NO heavy models!)
openai==1.54.0
pinecone-client==3.0.0

# Configuration and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Total size: ~50MB (all API clients, no ML models!)
# Perfect for serverless deployment on Render free tier
# Only 1 API key needed: OpenAI (for both embeddings and LLM)
